pyproject.toml
[tool.poetry]
name = "pls-blockchain"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = [{include = "blockchain"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
pycryptodome = "^3.16.0"
requests = "^2.28.1"
black = "^23.1.0"
pytest = "^7.2.1"
pyyaml = "^6.0"
schedule = "^1.1.0"
pkcs7 = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
